refactor: Open Graph 구현
